+++ CHANGELOG.md
@@ -59,6 +59,7 @@ release points are not being annotated in GitHub.
 - SIDD 2.0+ FilterType handling
 - Erroneous SIDD consistency error re: NITF NBPP when PixelType=RGB24I
 - Properly recompute SCPCOA metadata when updating SCP using a DEM in `sarpy.io.complex.converter.conversion_utility`
+- SIDD GeoInfo namespace handling
 
 ## [1.3.58] - 2023-08-07
 ### Added